+++ gems/aws-sdk-applicationinsights/lib/aws-sdk-applicationinsights.rb
@@ -52,6 +52,6 @@ require_relative 'aws-sdk-applicationinsights/customizations'
 # @!group service
 module Aws::ApplicationInsights
 
-  GEM_VERSION = '1.36.0'
+  GEM_VERSION = '1.37.0'
 
 end