+++ gems/aws-sdk-cloudhsmv2/lib/aws-sdk-cloudhsmv2.rb
@@ -52,6 +52,6 @@ require_relative 'aws-sdk-cloudhsmv2/customizations'
 # @!group service
 module Aws::CloudHSMV2
 
-  GEM_VERSION = '1.47.0'
+  GEM_VERSION = '1.48.0'
 
 end