+++ gems/aws-sdk-codestarnotifications/lib/aws-sdk-codestarnotifications.rb
@@ -52,6 +52,6 @@ require_relative 'aws-sdk-codestarnotifications/customizations'
 # @!group service
 module Aws::CodeStarNotifications
 
-  GEM_VERSION = '1.25.0'
+  GEM_VERSION = '1.26.0'
 
 end