+++ gems/aws-sdk-costexplorer/lib/aws-sdk-costexplorer.rb
@@ -52,6 +52,6 @@ require_relative 'aws-sdk-costexplorer/customizations'
 # @!group service
 module Aws::CostExplorer
 
-  GEM_VERSION = '1.86.0'
+  GEM_VERSION = '1.87.0'
 
 end