+++ gems/aws-sdk-pipes/lib/aws-sdk-pipes.rb
@@ -52,6 +52,6 @@ require_relative 'aws-sdk-pipes/customizations'
 # @!group service
 module Aws::Pipes
 
-  GEM_VERSION = '1.6.0'
+  GEM_VERSION = '1.7.0'
 
 end