+++ gems/aws-sdk-route53resolver/lib/aws-sdk-route53resolver.rb
@@ -52,6 +52,6 @@ require_relative 'aws-sdk-route53resolver/customizations'
 # @!group service
 module Aws::Route53Resolver
 
-  GEM_VERSION = '1.44.0'
+  GEM_VERSION = '1.45.0'
 
 end